CR 6677459: incorrect processing of menu hotkeys
Special patch for worng key processing by WTK (CR 6658788)
introduced also the invalid processing of key event when
menu is open. The fix to the patch :) is to check whether
menu is open do nothing in this case. Thes changes should
be removed in later release together with the patch itself.

Reviewed by Olga Milovskaya, thanks.
No branch for the changes were created, merged from trunk, r10526:10527